+++ CotEditor/Sources/WindowPaneController.swift
@@ -29,6 +29,9 @@ final class WindowPaneController: NSViewController {
     
     // MARK: Private Properties
     
+    private lazy var titleForRespectSystemSetting: String = self.tabbingOptionMenu!.items.first!.title
+    
+    @IBOutlet private weak var tabbingOptionMenu: NSMenu?
     @IBOutlet private weak var tabbingSupportCaution: NSTextField?
     
     @objc private dynamic var editorOpaque: Bool = (UserDefaults.standard[.windowAlpha] == 1.0)
@@ -45,6 +48,25 @@ final class WindowPaneController: NSViewController {
         if (NSApp.delegate as! AppDelegate).supportsWindowTabbing {
             self.tabbingSupportCaution!.removeFromSuperview()
         }
+    }
+    
+    
+    override func viewWillAppear() {
+        
+        super.viewWillAppear()
+        
+        guard #available(macOS 10.12, *) else { return }
+        
+        // display the current system-wide user setting for window tabbing in "Respect System Setting" menu item.
+        let menu = self.tabbingOptionMenu!
+        let systemSettingLabel = menu.item(withTag: NSWindow.userTabbingPreference.rawValue)!.title
+        let attrLabel = NSAttributedString(string: self.titleForRespectSystemSetting,
+                                           attributes: [.font: menu.font])
+        let userSettingLabel = NSAttributedString(string: String(format: NSLocalizedString(" (%@)", comment: ""), systemSettingLabel),
+                                                  attributes: [.font: menu.font,
+                                                               .foregroundColor: NSColor.secondaryLabelColor])
+        
+        menu.items.first!.attributedTitle = attrLabel + userSettingLabel
     }