+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Change Log
 - Remove the window size setting option.
     - Now, the initial window size is determined by the last window size, just like other standard Cocoa applications.
 - Remove the initlal document inspector visibility setting option.
+- Display current user's system-wide setting for window tabbing in the menu on the Window pane. 
 - [trivial] Update editor opacity sample tip.